% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barkNO.R
\name{barkNO}
\alias{barkNO}
\title{Bark model for spruce, pine, and birch for Norway}
\usage{
barkNO(d, h, dbh, h_top, sp = "spruce")
}
\arguments{
\item{d}{diameter (cm)}

\item{h}{height (m)}

\item{dbh}{diameter at breast height (cm)}

\item{h_top}{tree height (m)}

\item{sp}{species ('spruce','pine' or 'birch'; 1:3)}
}
\value{
double bark thickness at h (mm).
}
\description{
The bark model is based on Gordon (1983) via Stängle (2016).
}
\examples{
barkNO(d=30, h=1,dbh=25,h_top=30,sp="pine")

taper<-taperNO(h=(0:300)/10,dbh=25,h_top=30,sp="pine")
bark<-barkNO(d=taper, h=(0:300)/10,dbh=25,h_top=30,sp="pine")
plot((0:300)/10,taper,type="l",col="brown",xlab="height (m)",ylab="taper/bark (cm)")
points((0:300)/10,taper-bark/10,type="l",col="green")

plot((0:300)/10,bark,type="l",ylab="bark (mm)",xlab="height (m)")
}
