% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barkNOR.R
\name{barkNOR}
\alias{barkNOR}
\title{Bark model for spruce, pine, and birch for Norway}
\usage{
barkNOR(d, h, dbh, h_top, sp = "spruce")
}
\arguments{
\item{d}{diameter (cm).}

\item{h}{height (m).}

\item{dbh}{diameter at breast height (cm).}

\item{h_top}{tree height (m).}

\item{sp}{species ('spruce','pine' or 'birch'; 1:3).}
}
\value{
double bark thickness at h (mm).
}
\description{
The bark models are based on Gordon (1983) via Stängle (2016). Height is distance from stem base.
}
\examples{
barkNOR(d=30, h=1,dbh=25,h_top=30,sp="pine")

taper<-taperNOR(h=(0:300)/10,dbh=25,h_top=30,sp="pine")
bark<-barkNOR(d=taper, h=(0:300)/10,dbh=25,h_top=30,sp="pine")
plot((0:300)/10,taper,type="l",col="brown",xlab="height (m)",ylab="taper/bark (cm)")
points((0:300)/10,taper-bark/10,type="l",col="green")

plot((0:300)/10,bark,type="l",ylab="bark (mm)",xlab="height (m)")
}
