% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlocation.R
\name{dlocation}
\alias{dlocation}
\title{Estimate height of diameters along the stem}
\usage{
dlocation(dbh, h_top, d, sp = "spruce")
}
\arguments{
\item{dbh}{diameter at breast height (cm)}

\item{h_top}{height of diameter measurements (m)}

\item{d}{diameters (cm)}

\item{sp}{species}
}
\value{
Timber volume (m^3).
}
\description{
Estimates heights of given diameters using optimization of the taper function.
}
\examples{

dlocation(dbh=35,h_top=35,d=c(25,40,12,1,70),sp="birch")

h_vol<-dlocation(dbh=35,h_top=35,d=c(21,19))
volume(35,35,h_vol_lower = h_vol[1])
volume(35,35,h_vol_lower = h_vol[1],h_vol_upper=h_vol[2])

}
