% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfromd.R
\name{hfromd}
\alias{hfromd}
\title{Estimate tree height from diameters}
\usage{
hfromd(d, h, sp = "spruce", output = "H")
}
\arguments{
\item{d}{diameters}

\item{h}{height of diameters}

\item{sp}{species}

\item{output}{output from optimization: 'H' outputs the tree height, 'all' result of optim.}
}
\value{
Timber volume in m.
}
\description{
Estimates tree height from diameters using optimization of the taper function.
}
\examples{
volume(20, 30)
volume(c(20,25,30), c(30,25,37))
}
