% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kublin_no.R
\name{kublin_no}
\alias{kublin_no}
\title{Taper model for spruce, pine, and birch for Norway with dynamic diameter input}
\usage{
kublin_no(Hx, Hm, Dm, mHt, sp = 1)
}
\arguments{
\item{Hx}{height where to return diameters}

\item{Hm}{height of measured diameters}

\item{Dm}{measured diameters}

\item{mHt}{measured tree height}

\item{sp}{species}
}
\value{
Timber volume in m.
}
\description{
Allows the use of multiple diameters measured at various points along the stem to estimate taper.
}
\examples{

}
