% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taper.R
\name{plot_taper}
\alias{plot_taper}
\title{Plot taper curve}
\usage{
plot_taper(dbh, h_top, sp = "spruce")
}
\arguments{
\item{dbh}{diameter at breast height (cm).}

\item{h_top}{tree height (m).}

\item{sp}{species}
}
\description{
Plots the taper curve
}
\examples{
plot_taper(33,30)
}
