% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taperNO.R
\name{taperNO}
\alias{taperNO}
\title{Taper model for spruce, pine, and birch for Norway}
\usage{
taperNO(h, dbh, h_top, sp = "spruce")
}
\arguments{
\item{h}{heights where to return diameters (m)}

\item{dbh}{diameter at breast height (cm)}

\item{h_top}{tree height (m)}

\item{sp}{species ('spruce','pine' or 'birch'; 1:3)}
}
\value{
diameters at h (cm).
}
\description{
The taper model is based on Kozak 1988.
}
\examples{
taperNO(h=1:30,dbh=20,h_top=30,sp="pine")
}
