% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume}
\alias{volume}
\title{Calculate volume}
\usage{
volume(dbh, h_top, sp = "spruce")
}
\arguments{
\item{dbh}{numeric vector of diameters at breast height (cm).}

\item{h_top}{numeric vector of tree heights (m).}

\item{sp}{species}
}
\value{
Timber volume (m).
}
\description{
Calculate volume from taper model by integerating the taper function.
}
\examples{
volume(20, 30)
volume(c(20,25,30), c(30,25,37))
}
