% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barkNOR.R
\name{barkNOR}
\alias{barkNOR}
\title{Bark model for spruce, pine, and birch for Norway}
\usage{
barkNOR(d, h, dbh, h_top, sp = "spruce")
}
\arguments{
\item{d}{diameter (cm).}

\item{h}{height above ground (m).}

\item{dbh}{diameter at breast height, 1.3 m above ground (cm).}

\item{h_top}{tree height above ground (m).}

\item{sp}{species ('spruce','pine' or 'birch'; 1:3)}
}
\value{
double bark thickness (mm) at height h with diameter d.
}
\description{
The bark models are based on Hannrup B. 2004. Funktioner för skattning av barkens tjocklek hos tall och gran vid avverkning med skördare. (Functions for prediction of bark thickness of Norway spruce and Scots pine at CTL-harvesting). Arbetsrapport 575. Skogforsk, Uppsala.
via Stängle et al. 2017. Comparison of models for estimating bark thickness of Picea abies in southwest Germany: the role of tree, stand, and environmental factors. Ann For Sci. 74(1):16.
}
\examples{
barkNOR(d=30, h=1,dbh=25,h_top=30,sp="pine")

taper<-taperNOR(h=(0:300)/10,dbh=25,h_top=30,sp="pine")
bark<-barkNOR(d=taper, h=(0:300)/10,dbh=25,h_top=30,sp="pine")
plot((0:300)/10,taper,type="l",col="brown",xlab="height (m)",ylab="taper/bark (cm)")
points((0:300)/10,taper-bark/10,type="l",col="green")

plot((0:300)/10,bark,type="l",ylab="bark (mm)",xlab="height (m)")
}
