% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlocation.R
\name{dlocation}
\alias{dlocation}
\title{Estimate height of diameters along the stem}
\usage{
dlocation(dbh, h_top, d, sp = "spruce", with_bark = TRUE)
}
\arguments{
\item{dbh}{diameter at breast height over bark (cm)}

\item{h_top}{height of diameter measurements (m)}

\item{d}{diameters (cm)}

\item{sp}{species}

\item{with_bark}{diameter with (TRUE, default) or without bark (FALSE).}
}
\value{
Timber volume (m^3).
}
\description{
Estimates heights of given diameters along the stem using optimization of the taper function.
}
\examples{

dlocation(dbh=35,h_top=35,d=c(25,40,12,2,70),sp="birch")
dlocation(dbh=35,h_top=35,d=c(25,40,12,2,70),sp="birch",with_bark=FALSE)

D<-35
H<-34
T<-"G"

#Case 1:
dia<-21
h_vol<-dlocation(d=dia,dbh=D,h_top=H,sp=T)
volume(D,H,sp=T,h_vol_lower = h_vol)

#Case2:
dia<- c(21,16)
A<-1.2
h_vol<- dlocation(d=dia,dbh=D,h_top=H,sp=T)
if (diff(h_vol)>A) volume(D,H,sp=T,h_vol_lower = h_vol[1],h_vol_upper=h_vol[2])

}
