% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfromd.R
\name{hfromd}
\alias{hfromd}
\title{Estimate tree height from diameters}
\usage{
hfromd(d, h, sp = "spruce", output = "h", grd_search = F)
}
\arguments{
\item{d}{diameters (cm).}

\item{h}{height of diameter measurements (m).}

\item{sp}{species}

\item{output}{output from optimization: 'h' outputs the tree height,'d' diameter at breast height, 'all' result of optim.}

\item{grd_search}{if TRUE optimization is run on a matrix of initial parameters. The result of the best fit is returned. Might give better result at cost of performance.}
}
\value{
Timber volume in m.
}
\description{
Estimates tree height from diameters using optimization of the taper function. Height is distance from stem base.
}
\examples{

hfromd(d=c(39,27),h=c(2,7),sp="birch")
}
