% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taper.R
\name{plot_taper}
\alias{plot_taper}
\title{Plot taper curve}
\usage{
plot_taper(dbh, h_top, sp = "spruce")
}
\arguments{
\item{dbh}{diameter at breast height (cm).}

\item{h_top}{tree height (m).}

\item{sp}{species}
}
\description{
Plots the taper curve
}
\examples{
plot_taper(33,30)
plot_taper(dbh=c(33,20,18),h_top=c(30,25,20),sp=c(1,1,3))
plot_taper(dbh=rep(25,3),h_top=rep(27,3),sp=1:3)
}
