% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taperNO.R
\name{taperNO}
\alias{taperNO}
\title{Taper model for spruce, pine, and birch for Norway}
\usage{
taperNO(h, dbh, h_top, sp = "spruce", with_bark = TRUE)
}
\arguments{
\item{h}{heights where to return diameters (m)}

\item{dbh}{diameter at breast height over bark (cm)}

\item{h_top}{tree height (m)}

\item{sp}{species ('spruce','pine' or 'birch'; 1:3)}

\item{with_bark}{estimate diameter over (TRUE, default) or under bark (FALSE)}
}
\value{
diameters at h (cm).
}
\description{
The taper model is based on Kozak 1988.
}
\examples{
taperNO(h=1:30,dbh=20,h_top=30,sp="pine",with_bark=TRUE)
taperNO(h=1:30,dbh=20,h_top=30,sp="pine",with_bark=FALSE)
}
