% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taperNO.R
\name{taperNO}
\alias{taperNO}
\title{Taper model for spruce, pine, and birch for Norway}
\usage{
taperNO(h, dbh, h_top, sp = "spruce")
}
\arguments{
\item{h}{heights where to return diameters}

\item{dbh}{diameter at breast height in cm}

\item{h_top}{tree height in m}

\item{sp}{species}
}
\value{
Timber volume in m.
}
\description{
The taper model is based on Kozak 1988.
}
\examples{
taperNO(h=1:10,dbh=20,h_top=30,sp="pine")
}
