% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume}
\alias{volume}
\title{Calculate volume}
\usage{
volume(
  dbh,
  h_top,
  h_vol_lower = 0,
  h_vol_upper = NA,
  sp = "spruce",
  with_bark = TRUE
)
}
\arguments{
\item{dbh}{numeric vector of diameters at breast height (cm).}

\item{h_top}{numeric vector of tree heights (m).}

\item{h_vol_lower, h_vol_upper}{numeric vectors of heights between stem volume is calculated (m). NA in h_vol_upper is replaced by h_top.}

\item{sp}{species.}

\item{with_bark}{calculate volume over (TRUE, default) or under bark (FALSE).}
}
\value{
Timber volume (m^3).
}
\description{
Calculate volume from taper model by integerating the taper function. Height is distance from stem base.
}
\examples{
volume(20, 30)
volume(dbh=c(20,25,30), h_top=c(30,25,37))
volume(dbh=rep(25,11), h_top=rep(30,11),h_vol_lower=seq(0,30,3))
volume(dbh=rep(25,11), h_top=rep(30,11),h_vol_upper=seq(0,30,3))
volume(dbh=rep(25,11), h_top=rep(30,11),h_vol_upper=seq(0,30,3),with_bark=FALSE)
}
